+++ FBSDKCoreKit/FBSDKCoreKit/include/FBSDKAuthenticationToken.h
@@ -8,6 +8,8 @@
 
 #import <Foundation/Foundation.h>
 
+#import <FBSDKCoreKit/FBSDKAuthenticationTokenProtocols.h>
+
 @class FBSDKAuthenticationTokenClaims;
 @protocol FBSDKTokenCaching;
 
@@ -15,7 +17,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 /// Represent an AuthenticationToken used for a login attempt
 NS_SWIFT_NAME(AuthenticationToken)
-@interface FBSDKAuthenticationToken : NSObject <NSCopying, NSObject, NSSecureCoding>
+@interface FBSDKAuthenticationToken : NSObject <NSCopying, NSObject, NSSecureCoding, FBSDKAuthenticationTokenProviding, FBSDKAuthenticationTokenSetting>
 
 /**
  Internal init method exposed to facilitate transition to Swift.
@@ -36,7 +38,7 @@ NS_SWIFT_NAME(AuthenticationToken)
  The `currentAuthenticationToken` represents the authentication token of the
  current user and can be used by a client to verify an authentication attempt.
  */
-@property (class, nullable, nonatomic, copy) FBSDKAuthenticationToken *currentAuthenticationToken;
+@property (class, nullable, nonatomic, copy) FBSDKAuthenticationToken *currentAuthenticationToken NS_SWIFT_NAME(current);
 
 /// The raw token string from the authentication response
 @property (nonatomic, readonly, copy) NSString *tokenString;