+++ TestTools/TestTools/TestAuthenticationTokenWallet.swift
@@ -12,10 +12,10 @@ import Foundation
 @objcMembers
 public final class TestAuthenticationTokenWallet: NSObject, AuthenticationTokenProviding, AuthenticationTokenSetting {
   public static var tokenCache: TokenCaching?
-  public static var currentAuthenticationToken: AuthenticationToken?
+  public static var current: AuthenticationToken?
 
   public static func reset() {
     tokenCache = nil
-    currentAuthenticationToken = nil
+    current = nil
   }
 }